Address PR #42 code review feedback
## Critical Fixes

1. **ParticlePool activeCount_ double-decrement** (ParticlePool.h)
   - Fixed bug where particles killed via pool_.kill() decremented activeCount_ twice
   - Added explicit kill detection: only auto-decrement for natural death
   - kill() sets both intensity=0 AND age=255 as signature
   - updateAll() now distinguishes explicit kills from natural expiration

2. **FireParticle integer overflow in color calculations** (FireParticle.cpp)
   - Fixed uint8_t multiplication overflow before cast to uint32_t
   - Changed: ((uint32_t)(intensity * 3)) → (((uint32_t)intensity * 3))
   - Prevents color corruption at high intensity values

3. **LightningParticle double-fade** (LightningParticle.cpp)
   - Removed FADE flag from spawned particles
   - Kept manual fade in updateParticle() for precise control
   - Eliminated conflict between auto-fade and manual fade

4. **Additive blending integer overflow** (WaterParticle.cpp, FireParticle.cpp)
   - Fixed uint8_t addition overflow in min() calls
   - Changed: min(255, existing.r + r) → min(255, (int)existing.r + r)
   - Prevents color clamping errors in bright scenes

5. **Mass validation** (ParticlePool.h)
   - Added mass clamping to prevent division by zero
   - Minimum mass: 0.01f (prevents RadialForce crash)

## Documentation Improvements

6. **Clarified delta-time units and velocity normalization**
   - Updated Particle.h: velocity units are LEDs/sec with 30 FPS normalization
   - Updated ParticleGenerator.h: explained frame rate normalization factor
   - Updated ParticleForce.h: clarified force units

## Testing
- ✅ Compiles successfully (183,360 bytes, 22% flash)
- ✅ All static analysis warnings resolved

+++ blinky-things/generators/FireParticle.cpp
@@ -120,23 +120,23 @@ void FireParticle::renderParticle(const Particle* p, PixelMatrix& matrix) {
         uint8_t g = (color >> 8) & 0xFF;
         uint8_t b = color & 0xFF;
 
-        // Additive blending with existing heat
+        // Additive blending with existing heat (cast to int to prevent overflow)
         RGB existing = matrix.getPixel(x, y);
         matrix.setPixel(x, y,
-                       min(255, existing.r + r),
-                       min(255, existing.g + g),
-                       min(255, existing.b + b));
+                       min(255, (int)existing.r + r),
+                       min(255, (int)existing.g + g),
+                       min(255, (int)existing.b + b));
     }
 }
 
 uint32_t FireParticle::particleColor(uint8_t intensity) const {
     // Fire palette: black -> red -> orange -> yellow
     if (intensity < 85) {
-        // Black to red
-        return ((uint32_t)(intensity * 3) << 16);
+        // Black to red (cast before multiply to prevent overflow)
+        return (((uint32_t)intensity * 3) << 16);
     } else if (intensity < 170) {
-        // Red to orange
-        uint8_t green = (intensity - 85) * 3;
+        // Red to orange (cast before multiply to prevent overflow)
+        uint8_t green = (uint8_t)(((uint32_t)(intensity - 85)) * 3);
         return (0xFF0000 | ((uint32_t)green << 8));
     } else {
         // Orange to yellow
